Move MessageCollector into its own file
